organise layers of dcm
