+++ layers.yml
@@ -20,6 +20,7 @@ Internal Levelled Modules:
         - pymedphys.utilities
         - pymedphys.xarray
     layers:
+        - (_level4)
         - (_level3)
         - (_level2)
         - _level1
