+++ src/pymedphys/dcm/__init__.py
@@ -45,14 +45,15 @@ Available Functions
 
 from ..libutils import clean_and_verify_levelled_modules
 
-from ._level1.dcmdose import *
 from ._level1.dcmcreate import *
 from ._level2.header_tweaks import *
 from ._level2.dcmanonymise import *
 from ._level2.dcmstruct import *
 from ._level3.pydicom_wrapper import *
+from ._level4.dcmdose import *
 
 clean_and_verify_levelled_modules(globals(), [
-    '._level1.dcmdose', '._level1.dcmcreate', '._level2.header_tweaks',
-    '._level2.dcmanonymise', '._level2.dcmstruct', '._level3.pydicom_wrapper'
+    '._level1.dcmcreate', '._level2.header_tweaks',
+    '._level2.dcmanonymise', '._level2.dcmstruct', '._level3.pydicom_wrapper',
+    '._level4.dcmdose'
 ], package='pymedphys.dcm')
