+++ src/pymedphys/dcm/_level2/dcmstruct.py
@@ -32,13 +32,71 @@ from ...libutils import get_imports
 from ...geometry import (
     cubify_cube_definition, cube_vertices, cube_vectors)
 
-from .._level1.dcmdose import (
-    pull_structure, contour_to_points)
-
 IMPORTS = get_imports(globals())
 
 
 # pylint: disable=C0103
+
+def pull_structure_by_number(number, dcm_struct):
+    contours_by_slice_raw = [
+        item.ContourData
+        for item in dcm_struct.ROIContourSequence[number].ContourSequence
+    ]
+    x = [
+        np.array(item[0::3])
+        for item in contours_by_slice_raw]
+    y = [
+        np.array(item[1::3])
+        for item in contours_by_slice_raw]
+    z = [
+        np.array(item[2::3])
+        for item in contours_by_slice_raw]
+
+    return x, y, z
+
+
+def pull_structure(string, dcm_struct):
+    structure_names = np.array(
+        [item.ROIName for item in dcm_struct.StructureSetROISequence])
+    reference = structure_names == string
+    if np.all(reference == False):  # pylint: disable=C0121
+        raise Exception("Structure not found (case sensitive)")
+
+    index = int(np.where(reference)[0])
+    x, y, z = pull_structure_by_number(
+        index, dcm_struct)
+
+    return x, y, z
+
+
+def list_structures(dcm_struct):
+    return [item.ROIName for item in dcm_struct.StructureSetROISequence]
+
+
+def resample_contour(contour, n=50):
+    tck, u = splprep(
+        [contour[0], contour[1], contour[2]], s=0, k=1)
+    new_points = splev(np.arange(0, 1, 1/n), tck)
+
+    return new_points
+
+
+def resample_contour_set(contours, n=50):
+
+    resampled_contours = [
+        resample_contour([x, y, z], n)
+        for x, y, z in zip(*contours)
+    ]
+
+    return resampled_contours
+
+
+def contour_to_points(contours):
+    resampled_contours = resample_contour_set([
+        contours[1], contours[0], contours[2]])
+    contour_points = np.concatenate(resampled_contours, axis=1)
+
+    return contour_points
 
 
 def get_structure_aligned_cube(x0: np.ndarray, structure_name: str,
