+++ src/pymedphys/dcm/_level4/dcmdose.py
@@ -39,6 +39,10 @@ import pydicom
 import pydicom.uid
 
 from ...libutils import get_imports
+
+from .._level2.dcmstruct import pull_structure
+from .._level3.pydicom_wrapper import Dicom
+
 IMPORTS = get_imports(globals())
 
 Coords = namedtuple('Coords', 'x y z')
@@ -46,6 +50,10 @@ Dose = namedtuple(
     'Dose', 'values units type summation heterogeneity_correction')
 
 # pylint: disable=C0103
+
+
+class DicomDose(Dicom):
+    pass
 
 
 def load_dose_from_dicom(dcm, set_transfer_syntax_uid=True, reshape=True):
@@ -610,38 +618,6 @@ def average_bounding_profiles(dcm, depth_adjust, depth_lookup,
         return inplane, inplane_dose, crossplane, crossplane_dose
 
 
-def pull_structure_by_number(number, dcm_struct):
-    contours_by_slice_raw = [
-        item.ContourData
-        for item in dcm_struct.ROIContourSequence[number].ContourSequence
-    ]
-    x = [
-        np.array(item[0::3])
-        for item in contours_by_slice_raw]
-    y = [
-        np.array(item[1::3])
-        for item in contours_by_slice_raw]
-    z = [
-        np.array(item[2::3])
-        for item in contours_by_slice_raw]
-
-    return x, y, z
-
-
-def pull_structure(string, dcm_struct):
-    structure_names = np.array(
-        [item.ROIName for item in dcm_struct.StructureSetROISequence])
-    reference = structure_names == string
-    if np.all(reference == False):  # pylint: disable=C0121
-        raise Exception("Structure not found (case sensitive)")
-
-    index = int(np.where(reference)[0])
-    x, y, z = pull_structure_by_number(
-        index, dcm_struct)
-
-    return x, y, z
-
-
 def _get_index(z_list, z_val):
     indices = np.array([item[0] for item in z_list])
     # This will error if more than one contour exists on a given slice
@@ -705,36 +681,6 @@ def create_dvh(structure, dcm_struct, dcm_dose):
     plt.title('DVH')
     plt.xlabel('Dose (Gy)')
     plt.ylabel('Relative Volume (%)')
-
-
-def list_structures(dcm_struct):
-    return [item.ROIName for item in dcm_struct.StructureSetROISequence]
-
-
-def resample_contour(contour, n=50):
-    tck, u = splprep(
-        [contour[0], contour[1], contour[2]], s=0, k=1)
-    new_points = splev(np.arange(0, 1, 1/n), tck)
-
-    return new_points
-
-
-def resample_contour_set(contours, n=50):
-
-    resampled_contours = [
-        resample_contour([x, y, z], n)
-        for x, y, z in zip(*contours)
-    ]
-
-    return resampled_contours
-
-
-def contour_to_points(contours):
-    resampled_contours = resample_contour_set([
-        contours[1], contours[0], contours[2]])
-    contour_points = np.concatenate(resampled_contours, axis=1)
-
-    return contour_points
 
 
 def check_dcmdose(dcm):
